visual: extract classes from vmenus.c to classes.c
