+++ librz/cons/meson.build
@@ -12,6 +12,7 @@ rz_cons_sources = [
   'output.c',
   'pager.c',
   'pal.c',
+  'prompt.c',
   'cpipe.c',
   'rgb.c',
   'cutf8.c'
